cloud_functions/phase_05_classify/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
open3d==0.18.*
numpy==1.*
scikit-learn==1.*
